Make Tiger Turn Towards Foe During Charge
Even though it would be cool if the Tiger doesn't face the foe while charging up, I need this for consistency. Whatever I changed, I needed this for consistency. attackRotation is determined with regular attack, so there was a good chance nothing was determing attackRotation when I first tried targeted TigerSpecial. The weird part is that the Tiger still rotated the right direction. Neeevermind. The Tiger never rotated the right way to turn to the foe. I'm happy about that because that means this was important and that means there's nothing I don't understand. No logic breaks for me.